--- twig/Infobokser/avlys.html.twig
@@ -1,17 +1,0 @@
-<div class="row">
-	<div class="col-xs-12 col-sm-12">
-		<h4 class="mt-4">Korona-status 😷</h4>
-		<ol class="list-group" style="margin-bottom: .4em;">
-			<li class="list-group-item">
-				Hvis arrangementet må avlyses, eller arrangementet du skal sende videre til er avlyst,
-                eller om du har annen viktig informasjon, kan du legge inn dette her.
-				<div class="clearfix"></div>
-			</li>
-		</ol>
-		<div class="mt-0" align="right">
-        	<a href="#" class="goTo btn btn-ukm btn-xs btn-primary" data-action="avlys">rediger korona-informasjon</a>
-		</div>
-		<div class="clearfix"></div>
-	</div>
-</div>
-<div class="clearfix"></div>
