+++ twig/Monstring/kontaktpersoner.html.twig
@@ -1,4 +1,4 @@
-<div class="row" id="container_kontaktpersoner">
+<div class="row hvisSynlig" style="{{ display_hvisSynlig }}">
     <div class="col-12 col-sm-12">
         <h4 style="margin-top: 2em;">Kontaktpersoner</h4>
