+++ twig/Monstring/pamelding_workshop.html.twig
@@ -18,9 +18,10 @@
 					</div>
 					<div class="col-xs-12 col-sm-5" align="right">
 						<div class="input-group" style="margin-bottom: 1em;">
-							<input class="datepicker_kommune form-control" type="text" name="frist_1" value="{{ arrangement.getFrist1()|dato('d.m.Y') }}"/>
+							<input class="datepicker_kommune form-control" type="text" name="pl_deadline" value="{{ arrangement.getFrist1()|dato('d.m.Y') }}"/>
 							<span class="input-group-addon">kl. 23:59</span>
 						</div>
+						<input type="hidden" name="pl_deadline_time" value="23:59" />
 					</div>
 				</div>
 			</li>
@@ -44,7 +45,7 @@
 				</div>
 			</li>
 		</ol>
-		<div id="harDeltakerskjemaDetaljer" style="display:none;" align="right">
+		<div id="harDeltakerskjemaDetaljer" {{ not arrangement.harDeltakerskjema() ? 'style="display:none;"' }} align="right">
 			<div class="pull-right">
 				<a href="#" class="goTo btn btn-ukm btn-xs btn-primary" data-action="personskjema" style="margin-top: -.8em;">rediger skjema</a>
 			</div>
