+++ twig/Videresending/basics.html.twig
@@ -1,0 +1,48 @@
+{# VIDERESENDING #}
+<ol class="list-group-item">
+	<li class="list-group-item">
+		<div class="row" style="margin-left:-30px; margin-right: -30px;">
+			<div class="col-xs-12 col-sm-6">
+				<div class="col-xs-12 col-sm-12">
+					<label>Når åpner videresendingen?</label>
+				</div>
+				<div class="col-xs-7 col-sm-6">
+					<div class="form-group">
+						<input class="datepicker_{{ arrangement.getType() }} form-control" type="text" name="videresending_start" value="{{ arrangement.getVideresendingApner()|dato('d.m.Y') }}">
+					</div>
+				</div>
+				<div class="col-xs-5 col-sm-6">
+					<div class="input-group">
+						<span class="input-group-addon">kl.
+						</span>
+						<select class="form-control" name="videresending_start_time">
+							{% include "hourPicker.html.twig" with { current: arrangement.getVideresendingApner() } %}
+						</select>
+					</div>
+				</div>
+			</div>
+
+			<div class="col-xs-12 col-sm-6">
+				<div class="col-xs-12 col-sm-12">
+					<label>Når stenger videresendingen?</label>
+				</div>
+				<div class="col-xs-7 col-sm-6">
+					<div class="form-group">
+						<input class="datepicker_{{ arrangement.getType() }} form-control" type="text" name="videresending_start" value="{{ arrangement.getVideresendingStenger()|dato('d.m.Y') }}">
+					</div>
+				</div>
+				<div class="col-xs-5 col-sm-6">
+					<div class="input-group">
+						<span class="input-group-addon">kl.
+						</span>
+						<select class="form-control" name="videresending_start_time">
+							{% include "hourPicker.html.twig" with { current: arrangement.getVideresendingStenger() } %}
+						</select>
+					</div>
+				</div>
+			</div>
+		</div>
+
+		<div class="clearfix"></div>
+	</li>
+</ol>
