--- twig/vent-til-ny-sesong.html.twig
@@ -1,16 +1,0 @@
-<div class="container-fluid">
-    <div class="row">
-        <div class="col-12">
-            <h2>Endring av arrangement er ikke mulig</h2>
-
-            <p class="lead">
-                Mens vi gjør systemet klart for registrering av {{ now|date('Y')+1}}-sesongen, 
-                er endring av arrangement ikke mulig.
-                <br />
-                Så fort systemet er klart, sender UKM Norge ut nye passord til alle.
-            </p>
-
-            <div class="clearfix"></div>
-        </div>
-    </div>
-</div>
