+++ twig/videresending.html.twig
@@ -1,74 +1,19 @@
-	<div class="row" style="margin-top: 3em;">
-		<div class="col-12 col-sm-12">
-			<h4 class="mt-0">Videresending</h4>
+<h2>Videresending til
+	{{ arrangement.getNavn() }}</h2>
 
-			<p style="margin-top: 0.8em;">
-				{% if arrangement.harPamelding() %}
-					I tillegg til direkte påmelding, kan du også ta i mot videresendte fra andre arrangementer.
-				{% endif %}
+<p>
+	{% if arrangement.harPamelding() %}
+		I tillegg til direkte påmelding, kan du også ta i mot videresendte fra andre arrangementer.
+	{% else %}
+		Fordi arrangementet ikke har direkte påmelding, kommer alle deltakerne som videresending fra andre arrangementer.
+	{% endif %}
+</p>
 
-				Fristen for å videresende er den samme som påmeldingsfristene.
-			</p>
-			{% if not arrangement.harPamelding() %}
-				<p>
-					Fordi arrangementet ikke har direkte påmelding, kommer alle deltakerne som videresending fra andre arrangementer.
-				</p>
-			{% endif %}
 
-			{# VIDERESENDING #}
-			<ol class="list-group-item">
-				<li class="list-group-item" id="videresending_frister_selector">
-					<div class="row" style="margin-left:-30px; margin-right: -30px;">
-						<div class="col-xs-12 col-sm-6">
-							<div class="col-xs-12 col-sm-12">
-								<label>Når åpner videresendingen?</label>
-							</div>
-							<div class="col-xs-7 col-sm-6">
-								<div class="form-group">
-									<input class="datepicker_{{ arrangement.getType() }} form-control" type="text" name="videresending_start" value="{{ arrangement.getVideresendingApner()|dato('d.m.Y') }}">
-								</div>
-							</div>
-							<div class="col-xs-5 col-sm-6">
-								<div class="input-group">
-									<span class="input-group-addon">kl.
-									</span>
-									<select class="form-control" name="videresending_start_time">
-										{% include "hourPicker.html.twig" with { current: arrangement.getVideresendingApner() } %}
-									</select>
-								</div>
-							</div>
-						</div>
-
-						<div class="col-xs-12 col-sm-6">
-							<div class="col-xs-12 col-sm-12">
-								<label>Når stenger videresendingen?</label>
-							</div>
-							<div class="col-xs-7 col-sm-6">
-								<div class="form-group">
-									<input class="datepicker_{{ arrangement.getType() }} form-control" type="text" name="videresending_start" value="{{ arrangement.getVideresendingStenger()|dato('d.m.Y') }}">
-								</div>
-							</div>
-							<div class="col-xs-5 col-sm-6">
-								<div class="input-group">
-									<span class="input-group-addon">kl.
-									</span>
-									<select class="form-control" name="videresending_start_time">
-										{% include "hourPicker.html.twig" with { current: arrangement.getVideresendingStenger() } %}
-									</select>
-								</div>
-							</div>
-						</div>
-					</div>
-
-					<div class="clearfix"></div>
-				</li>
-			</ol>
-
-		</div>
-	</div>
-	{% include "Videresending/frist.html.twig" %}
-	{% include "Videresending/avsendere.html.twig" %}
-	{% include "Videresending/infotekst.html.twig" %}
-	{% include "Videresending/nominasjon.html.twig" %}
-	{% include "Videresending/ledere.html.twig" %}
-	{% include "Videresending/skjema.html.twig" %}
+{#% include "Videresending/basics.html.twig" %}
+{% include "Videresending/frist.html.twig" %}
+{% include "Videresending/avsendere.html.twig" %}
+{% include "Videresending/infotekst.html.twig" %}
+{% include "Videresending/nominasjon.html.twig" %}
+{% include "Videresending/ledere.html.twig" %}
+{% include "Videresending/skjema.html.twig" %#}
